backend/requirements.txt
# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
emails==0.6
httpx==0.26.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Pydantic & Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# Kafka
aiokafka==0.10.0

# Utils
python-dotenv==1.0.0
tenacity==8.2.3
structlog==24.1.0
orjson==3.9.12

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.0
ruff==0.1.14
mypy==1.8.0
